WIP(legend): move default legend icon logic from series to legend
